Segment Tree Solutions
